requirements.txt
imageio==1.5
numpy==1.10.2
Pillow==3.1.0
pycairo==1.10.0
